Extract TripleAMenu.java
